fetch-data/requirements.txt
hdx-python-api==6.1.4
boto3
io
pandas
regex
requests